made sure symlink not attempted if already exists

+++ theme_sync/recipes/mount_cloud9.rb
@@ -23,4 +23,5 @@ end
 execute "create symlink to themes" do
   command "ln -s /srv/www/themes/ /srv/www/oliver/shared/themes"
   user "deploy"
+  not_if do ::File.exists?('/srv/www/oliver/shared/themes/eventbeat/layout/theme.liquid') end
 end
